CCMSPUI-381 Checkstyle issues resolved
